% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rewbla_banding_data}
\alias{rewbla_banding_data}
\alias{rewbla_observations}
\alias{rewbla_intervals}
\title{Example banding data for Red-wing Blackbird}
\format{
An object of class \code{data.frame} with 6000 rows and 6 columns.
}
\usage{
rewbla_observations

rewbla_intervals
}
\description{
This data was extracted from the North American Bird Banding Program Dataset
1960-2022:

Celis-Murillo, A., Malorodova, M., and Nakash, E., 2022, North American Bird
Banding Program Dataset 1960-2022 retrieved 2022-07-14: U.S. Geological
Survey data release, https://doi.org/10.5066/P9BSM38F.

It has been filtered and randomly subset to 3000 birds each of which occurs
twice in observations and once in intervals. The sub-setting was for
demonstration not statistical reasons.
}
\keyword{datasets}
