% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\docType{data}
\name{states}
\alias{states}
\title{Low Resolution State Boundaries}
\format{
\emph{sf} object containing very coarse state boundaries this is also of
class \strong{data.frame} and has columns:
\describe{
\item{name}{State name}
\item{geometry}{Contains polyon information}

The code below was used to create this object:

\if{html}{\out{<div class="sourceCode">}}\preformatted{bf <- amewoo
states <- rnaturalearth::ne_states("United States of America")
states <- states[!states$gn_name \%in\% c("Alaska", "Hawaii"), ]
states <- sf::st_transform(states, crs(bf))


states <- sf::st_simplify(states, preserveTopology = FALSE,
                         dTolerance = 1000)
states <- states[ , names(states) \%in\% c("name", "geometry")]
}\if{html}{\out{</div>}}

}
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
states
}
\description{
This is a low resolution copy of the boundaries of U.S. states
from Natural Earth
}
\details{
It is included here to make it available to a vignette in BirdFlowR
without adding a dependency to the \strong{rnaturalearthhires} package.

If you need state boundaries for anything other than a trivial example
please use \code{rnaturalearth::ne_states("United States of America")} for
a higher resolution version of the state boundaries.
}
\keyword{datasets}
