% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BirdFlowModels}
\alias{BirdFlowModels}
\alias{amewoo}
\title{Birdflow Example Models}
\format{
A BirdFlow model
}
\source{
Trained on eBird Status and Trends 2019 Data.
}
\usage{
amewoo
}
\description{
This BirdFlow model is intended for for testing code and
demonstrating usage in \pkg{BirdFlowR}; it is not fully vetted and should
not be used for research or inference.

Currently there is just one model "amewoo" (American woodcock) which lacks
complete metadata. It will likely be replaced soon.
}
\examples{
print(amewoo)
species(amewoo)
}
\references{
Fink, D., T. Auer, A. Johnston, M. Strimas-Mackey, S. Ligocki, O.
Robinson, W. Hochachka, L. Jaromczyk, A. Rodewald, C. Wood, I. Davies, A.
Spencer. 2022. eBird Status and Trends, Data Version: 2021; Released: 2022.
Cornell Lab of Ornithology, Ithaca, New York.
https://doi.org/10.2173/ebirdst.2021
}
\keyword{datasets}
